+++ modules/pybind11/src/bindings/deserialize/tof_info_v3.h
@@ -89,7 +89,7 @@ bind_struct_tofinfov3(pybind11::module_& m)
         Parameters for calibration
       )");
 
-  py::class_<ifm3d::TofInfoV3, ifm3d::TofInfoV3::Ptr> tof_info_v3(
+  py::class_<ifm3d::TOFInfoV3, ifm3d::TOFInfoV3::Ptr> tof_info_v3(
     m,
     "ToFInfoV3",
     R"(
@@ -104,82 +104,82 @@ bind_struct_tofinfov3(pybind11::module_& m)
 
   tof_info_v3.def_readonly(
     "version",
-    &ifm3d::TofInfoV3::version,
+    &ifm3d::TOFInfoV3::version,
     R"(
          Version of the TOF_INFO data
       )");
 
   tof_info_v3.def_readonly(
     "amplitude_resolution",
-    &ifm3d::TofInfoV3::amplitude_resolution,
+    &ifm3d::TOFInfoV3::amplitude_resolution,
     R"(
          Resolution of the amplitude buffer
       )");
   tof_info_v3.def_readonly(
     "distance_resolution",
-    &ifm3d::TofInfoV3::distance_resolution,
+    &ifm3d::TOFInfoV3::distance_resolution,
     R"(
         Resolution of distance buffer per digit[m]
       )");
   tof_info_v3.def_readonly(
     "amp_normalization_factors",
-    &ifm3d::TofInfoV3::amp_normalization_factors,
+    &ifm3d::TOFInfoV3::amp_normalization_factors,
     R"(
           Amplitude normalization factors for the individual exposure times
       )");
   tof_info_v3.def_readonly(
     "extrisic_optic_to_user",
-    &ifm3d::TofInfoV3::extrisic_optic_to_user,
+    &ifm3d::TOFInfoV3::extrisic_optic_to_user,
     R"(
         Extrinsic optic parameter to user
       )");
   tof_info_v3.def_readonly(
     "intrinsic_calibration",
-    &ifm3d::TofInfoV3::intrinsic_calibration,
+    &ifm3d::TOFInfoV3::intrinsic_calibration,
     R"(
         Intrinsic calibration parameters
       )");
   tof_info_v3.def_readonly(
     "inverse_intrinsic_calibration",
-    &ifm3d::TofInfoV3::inverse_intrinsic_calibration,
+    &ifm3d::TOFInfoV3::inverse_intrinsic_calibration,
     R"(
         Inverse intrinsic calibration parameters
       )");
   tof_info_v3.def_readonly(
     "exposure_timestamps_ns",
-    &ifm3d::TofInfoV3::exposure_timestamps_ns,
+    &ifm3d::TOFInfoV3::exposure_timestamps_ns,
     R"(
          The timestamp of the individual exposure time [nano seconds]
       )");
   tof_info_v3.def_readonly(
     "exposure_times_s",
-    &ifm3d::TofInfoV3::exposure_times_s,
+    &ifm3d::TOFInfoV3::exposure_times_s,
     R"(
          Actual exposure times of a single phase image [seconds]
       )");
   tof_info_v3.def_readonly(
     "illu_temperature",
-    &ifm3d::TofInfoV3::illu_temperature,
+    &ifm3d::TOFInfoV3::illu_temperature,
     R"(
          Illumination temperature
       )");
   tof_info_v3.def_readonly(
     "mode",
-    &ifm3d::TofInfoV3::mode,
+    &ifm3d::TOFInfoV3::mode,
     R"(
         Mode of the head
       )");
   tof_info_v3.def_readonly(
     "imager",
-    &ifm3d::TofInfoV3::imager,
+    &ifm3d::TOFInfoV3::imager,
     R"(
         Imager type
       )");
 
   tof_info_v3.def_static(
     "Deserialize",
-    [](py::array_t<uint8_t, py::array::c_style | py::array::forcecast> in) -> ifm3d::TofInfoV3 {
-    ifm3d::TofInfoV3 val;
+    [](py::array_t<uint8_t, py::array::c_style | py::array::forcecast> in) -> ifm3d::TOFInfoV3 {
+    ifm3d::TOFInfoV3 val;
     val.Read(reinterpret_cast<const uint8_t*>(in.data(0)),in.nbytes());
     return val;
   },